+++ lib/critic.dart
@@ -21,16 +21,17 @@ class Critic {
   }
 
   String? _apiToken;
-  int? _appId;
+  String? _appId;
 
   Future<App> _createAppData() async {
     final PackageInfo info = await PackageInfo.fromPlatform();
     return App.create(
-        name: (info.appName.isEmpty) ? 'Unavailable' : info.appName,
-        package: info.packageName,
-        platform: Platform.isAndroid ? 'Android' : 'iOS',
-        versionName: info.version,
-        versionCode: info.buildNumber);
+      name: (info.appName.isEmpty) ? 'Unavailable' : info.appName,
+      package: info.packageName,
+      platform: Platform.isAndroid ? 'Android' : 'iOS',
+      versionName: info.version,
+      versionCode: info.buildNumber,
+    );
   }
 
   Future<Device> _createDeviceData() async {
@@ -38,38 +39,45 @@ class Critic {
     if (Platform.isAndroid) {
       AndroidDeviceInfo androidInfo = await deviceInfo.androidInfo;
       return Device(
-          identifier: androidInfo.id,
-          manufacturer: androidInfo.manufacturer,
-          model: androidInfo.model,
-          networkCarrier: 'Not available',
-          platform: 'Android',
-          platformVersion: androidInfo.version.release);
+        identifier: androidInfo.id,
+        manufacturer: androidInfo.manufacturer,
+        model: androidInfo.model,
+        networkCarrier: 'Not available',
+        platform: 'Android',
+        platformVersion: androidInfo.version.release,
+      );
     } else if (Platform.isIOS) {
       IosDeviceInfo iosInfo = await deviceInfo.iosInfo;
       return Device(
-          identifier: iosInfo.identifierForVendor ?? '',
-          manufacturer: 'Apple',
-          model: iosInfo.model,
-          networkCarrier: 'Not available',
-          platform: 'iOS',
-          platformVersion: iosInfo.systemVersion);
+        identifier: iosInfo.identifierForVendor ?? '',
+        manufacturer: 'Apple',
+        model: iosInfo.model,
+        networkCarrier: 'Not available',
+        platform: 'iOS',
+        platformVersion: iosInfo.systemVersion,
+      );
     }
     return Device(
-        identifier: 'unknown',
-        manufacturer: 'unknown',
-        model: 'unknown',
-        networkCarrier: 'Not available',
-        platform: 'Unknown',
-        platformVersion: Platform.version);
+      identifier: 'unknown',
+      manufacturer: 'unknown',
+      model: 'unknown',
+      networkCarrier: 'Not available',
+      platform: 'Unknown',
+      platformVersion: Platform.version,
+    );
   }
 
-  Future<bool> initialize(String apiToken) async {
+  Future<bool> initialize(String apiToken, {String? baseUrl}) async {
     _apiToken = apiToken;
+    if (baseUrl != null) {
+      Api.setBaseUrl(baseUrl);
+    }
 
     App appData = await _createAppData();
     Device deviceData = await _createDeviceData();
-    AppInstall response =
-        await Api.ping(PingRequest(apiToken: _apiToken!, app: appData, device: deviceData)).catchError((Object error) {
+    AppInstall response = await Api.ping(
+      PingRequest(apiToken: _apiToken!, app: appData, device: deviceData),
+    ).catchError((Object error) {
       print('Ping to critic failed: ' + error.toString());
       return Future<AppInstall>.error(false);
     });
@@ -78,9 +86,14 @@ class Critic {
   }
 
   Future<BugReport> submitReport(BugReport report) async {
-    assert(_apiToken != null, 'The API Token must be initialized using the initialize(String) call.');
-    assert(_appId != null,
-        'The App ID must be initialized. Make sure to call initialize(String). If you have done this, please check the logs to see why it failed.');
+    assert(
+      _apiToken != null,
+      'The API Token must be initialized using the initialize(String) call.',
+    );
+    assert(
+      _appId != null,
+      'The App ID must be initialized. Make sure to call initialize(String). If you have done this, please check the logs to see why it failed.',
+    );
     BugReportRequest requestData = BugReportRequest(
       appInstall: AppInstall(id: _appId!),
       apiToken: _apiToken!,