+++ lib/modal/report_request_modal.dart
@@ -6,11 +6,15 @@ class BugReportRequest {
   BugReport report;
   AppInstall appInstall;
 
-  BugReportRequest({required this.appInstall, required this.apiToken, required this.report});
+  BugReportRequest({
+    required this.appInstall,
+    required this.apiToken,
+    required this.report,
+  });
 
   Map<String, dynamic> toJson() => <String, dynamic>{
-        'api_token': apiToken,
-        'bug_report': report,
-        'app_install': appInstall,
-      };
+    'api_token': apiToken,
+    'bug_report': report,
+    'app_install': appInstall,
+  };
 }